Inject queue to view model for performing isButtonEnabled,  isLoading and labelText callbacks
This enables us to do the following:

- For production code:  we inject the main queue and we dispatch asynchronously  calling these callbacks to the main queue so that the ViewController can use them and bind data to views as is.

- For unit tests: we use 2 different ways to handle the asynchronous execution of these - for display purposes:
a. For the  labelText we are using the XCTExpectation mechanism
b. For the isButtonEnabled and isLoading callbacks, in the tests we are injecting a custom synchronous queue and use dispatching.sync {} after an asynchronous action in order to cause the test execution threat to wait for the asynchronous action.